+++ BE-ClothingSalesManager/src/main/java/com/example/beclothingsalesmanager/services/LoaiService.java
@@ -22,9 +22,18 @@ public class LoaiService {
         Loai loai = new Loai();
         loai.setTen(loaiRequest.getTen());
         loaiRepository.save(loai);
+
+        System.out.println("LoaiService.add: " + loai.getTen());
     }
 
     public void delete(Integer id) {
-        loaiRepository.deleteById(id);
+        Loai loai = loaiRepository.findById(id).orElse(null);
+        if (loai != null) {
+            loaiRepository.deleteById(id);
+
+            System.out.println("LoaiService.delete: " + id);
+        } else {
+            System.out.println("LoaiService.delete: null");
+        }
     }
 }
